requirements.txt
flask==3.0.3
scikit-learn==1.5.1
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
streamlit==1.37.0
requests==2.32.3
Pillow==10.1.0
matplotlib==3.9.1
seaborn==0.13.2
waitress==3.0.0
